refactor: Extract common functionality out of ConfigurationDocGenerator
