+++ docs/Configuration.md
@@ -231,14 +231,14 @@ Here is a list of all features which can be toggled on/off.
 
 | Feature                              | Description                                                                                                                                                                                       | Enabled by default | Sandbox |
 |--------------------------------------|---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:------------------:|:-------:|
-| `APIBikeRental`                      | Enable the bike rental endpoint.                                                                                                                                                                  |         ✓️         |         |
+| `APIBikeRental`                      | Enable the bike rental endpoint.                                                                                                                                                                  |                    |         |
 | `APIServerInfo`                      | Enable the server info endpoint.                                                                                                                                                                  |         ✓️         |         |
 | `APIGraphInspectorTile`              | Enable the inspector  endpoint for graph information for inspection/debugging purpose.                                                                                                            |         ✓️         |         |
 | `APIUpdaterStatus`                   | Enable endpoint for graph updaters status.                                                                                                                                                        |         ✓️         |         |
 | `ConsiderPatternsForDirectTransfers` | Enable limiting transfers so that there is only a single transfer to each pattern.                                                                                                                |         ✓️         |         |
 | `DebugClient`                        | Enable the debug web client located at the root of the web server.                                                                                                                                |         ✓️         |         |
 | `FloatingBike`                       | Enable floating bike routing.                                                                                                                                                                     |         ✓️         |         |
-| `MinimumTransferTimeIsDefinitive`    | If the minimum transfer time is a lower bound (default) or the definitive time for the transfer. Set this to `true` if you want to set a transfer time lower than what OTP derives from OSM data. |                    |         |
+| `MinimumTransferTimeIsDefinitive`    | If the minimum transfer time is a lower bound (default) or the definitive time for the transfer. Set this to `true` if you want to set a transfer time lower than what OTP derives from OSM data. |         ✓️         |         |
 | `OptimizeTransfers`                  | OTP will inspect all itineraries found and optimize where (which stops) the transfer will happen. Waiting time, priority and guaranteed transfers are taken into account.                         |         ✓️         |         |
 | `ParallelRouting`                    | Enable performing parts of the trip planning in parallel.                                                                                                                                         |                    |         |
 | `TransferConstraints`                | Enforce transfers to happen according to the _transfers.txt_(GTFS) and Interchanges(NeTEx). Turing this _off_ will increase the routing performance a little.                                     |         ✓️         |         |
