+++ src/test/java/org/opentripplanner/generate/doc/ConfigurationDocTest.java
@@ -1,30 +1,21 @@
 package org.opentripplanner.generate.doc;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.opentripplanner.framework.io.FileUtils.readFile;
+import static org.opentripplanner.framework.io.FileUtils.writeFile;
 import static org.opentripplanner.generate.doc.framework.ConfigTypeTable.configTypeTable;
 import static org.opentripplanner.generate.doc.framework.OTPFeatureTable.otpFeaturesTable;
 import static org.opentripplanner.generate.doc.framework.TemplateUtil.replaceSection;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.PrintWriter;
 import org.junit.jupiter.api.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-@SuppressWarnings("NewClassNamingConvention")
 public class ConfigurationDocTest {
-
-  private static final Logger LOG = LoggerFactory.getLogger(ConfigurationDocTest.class);
 
   private static final File FILE = new File("docs", "Configuration.md");
 
   private static final String CONFIG_TYPE_PLACEHOLDER = "CONFIGURATION-TYPES-TABLE";
   private static final String OTP_FEATURE_PLACEHOLDER = "OTP-FEATURE-TABLE";
-  public static final String NEW_LINE = "\n";
 
   /**
    * NOTE! This test updates the {@code docs/Configuration.md} document based on the latest
@@ -40,33 +31,11 @@ public class ConfigurationDocTest {
   @Test
   public void updateConfigurationDoc() {
     // Read and close inout file (same as output file)
-    String doc = readInConfigurationFile();
-    doc = replaceSection(doc, CONFIG_TYPE_PLACEHOLDER, air(configTypeTable()));
-    doc = replaceSection(doc, OTP_FEATURE_PLACEHOLDER, air(otpFeaturesTable()));
-    writeToConfigurationFile(doc);
+    String doc = readFile(FILE);
+    doc = replaceSection(doc, CONFIG_TYPE_PLACEHOLDER, configTypeTable());
+    doc = replaceSection(doc, OTP_FEATURE_PLACEHOLDER, otpFeaturesTable());
+    writeFile(FILE, doc);
 
-    assertEquals(doc, readInConfigurationFile());
-  }
-
-  private String readInConfigurationFile() {
-    try (var is = new FileInputStream(FILE)) {
-      return new String(is.readAllBytes(), UTF_8);
-    } catch (IOException e) {
-      throw new RuntimeException(e.getMessage(), e);
-    }
-  }
-
-  private void writeToConfigurationFile(String doc) {
-    try (var fileOut = new FileOutputStream(FILE)) {
-      var out = new PrintWriter(fileOut);
-      out.write(doc);
-      out.flush();
-    } catch (IOException e) {
-      throw new RuntimeException(e.getMessage(), e);
-    }
-  }
-
-  private static String air(String section) {
-    return NEW_LINE + NEW_LINE + section + NEW_LINE;
+    assertEquals(doc, readFile(FILE));
   }
 }
